fix(graphics): the clock renders at 30, and the rate owns the period
Sixty was a display's refresh, not anything the engine could hold: a
viewer already draws at MAX_VIEWER_FPS, which is 30, and a lossless
16-bit recording of a stage cannot be encoded at 60 on any machine
measured. The rate is now the constant and PERIOD derives from it,
rather than the rate being read back off a microsecond count.

+++ backend/graphics/goofi-graphics/src/runtime.rs
@@ -403,8 +403,8 @@ impl Runtime {
             if self.taping.contains_key(&uid) || rec.is_open(&id) {
                 continue;
             }
-            let kind = Kind::Video { size, fps: crate::FPS };
-            let live = match rec.open(&id, kind, t, StreamMeta::measured(Some(crate::FPS))) {
+            let kind = Kind::Video { size, fps: f64::from(crate::FPS) };
+            let live = match rec.open(&id, kind, t, StreamMeta::measured(Some(f64::from(crate::FPS)))) {
                 Ok(()) => true,
                 // A stream that will not open is this NODE's failure and nobody else's: the
                 // recording keeps every other stream, and the entry the recorder filed says why.
